Introduce microsecond timestamps to the specification
Given that the timestamp is already a 64-bit unsigned big-endian
integer, the spec will remain viable for over 500,000 years.

This allows us to more precisely compare token creation timestamps
against revocation event timestamps (which already have microsecond
precision). This helps us avoid a race condition wherein a token can be
created *after* a revocation event, but appear to be created *before*
it, thus matching the revocation event and being considered invalid.
After this change, the specification is still vulnerable to this race
condition, but the time window to reproduce it is narrowed to a single
microsecond, instead of a single second.

Closes #12